Add user info model and all users view
Added a new user info model to represent user data and implemented the all users view to display all users in the app. Also added a user info card widget to display user information in a card format. Removed the day by day table view as it was not being used.